Move MemoryPdfSplitter to a seperate file
